+++ src/TrashLib/Services/Sonarr/ISonarrGuideDataLister.cs
@@ -4,4 +4,5 @@ public interface ISonarrGuideDataLister
 {
     void ListReleaseProfiles();
     void ListTerms(string releaseProfileId);
+    void ListQualities();
 }