+++ src/TrashLib/Services/Sonarr/SonarrGuideDataLister.cs
@@ -1,6 +1,7 @@
 using System.Text;
 using CliFx.Infrastructure;
 using JetBrains.Annotations;
+using MoreLinq;
 using Serilog;
 using TrashLib.Services.Sonarr.ReleaseProfile;
 using TrashLib.Services.Sonarr.ReleaseProfile.Guide;
@@ -85,5 +86,18 @@ public class SonarrGuideDataLister : ISonarrGuideDataLister
         }
 
         _console.Output.WriteLine();
+    }
+
+    public void ListQualities()
+    {
+        _console.Output.WriteLine("\nList of Quality Definition types in the TRaSH Guides:\n");
+
+        _guide.GetQualities()
+            .Select(x => x.Type)
+            .ForEach(x => _console.Output.WriteLine($"  - {x}"));
+
+        _console.Output.WriteLine(
+            "\nThe above quality definition types can be used with the `quality_definition:` property in your " +
+            "recyclarr.yml file.");
     }
 }